moved block to the memory package.
